requirements.txt
gradio>=3.23.0
moviepy>=1.0.3
transparent-background==1.3.2
torch>=1.12.0
opencv-python>=4.5.5.62
Pillow>=9.2.0
numpy>=1.21.6
